Rough out property-based attributes and resources
The idea here is to make the JSON from Property objects representing the
values in a JSON object (or the object itself).  This should make things
like making a schema easier, because you can add things like a
description to individual properties, and the properties know their
type.

Also reworks relationships to be established using a picking method
where picks can be cross-referenced across relationships.  The basic
idea is to create all resources first, then use picking logic to figure
out how those resources should be associated.

This is all very messy, but it works!