Ensure attributes are stable for a given seed
Each of the sections (resource creation, relationship picking, and
attribute definiton) could call rand.  As a result the value picked for
an attribute can change if a relationship is added or removed.  That
screws up testing because the output values no longer are stable -- this
mitigates the problem for some use cases, although there's probably no
way to lock it down completely.